Move lib/invites/actions to state/invites/actions
